Remove diagnostic instrumentation, rename fenced-write helper
Strip the WF_TRACE replay tracing that was used to diagnose the
CORRUPTED_EVENT_LOG race \u2014 it's served its purpose now that the fix
is in. Specifically:

- Delete packages/core/src/__debug-replay-trace.ts and its 8 call sites
  in workflow.ts, step.ts, workflow/hook.ts, workflow/sleep.ts.
- Drop the matching [DEBUG] inline narrative comments at each call site.
- Rename packages/core/src/runtime/__fenced-write.ts \u2192 fenced-write.ts
  (the leading-underscore convention marked it as throwaway diagnostic
  code; the helper is intended to stay).
- Trim the file header on fenced-write.ts and the related narrative
  comment in suspension-handler.ts to drop the failing-runId / PR-number
  references that only made sense in the debug context.

No behavioral change. typecheck clean (0 errors); 1014/1014 unit tests
pass (same as parent commit 77f057a75).
